src/requirements.txt
psycopg2-binary
pangres
langchain
replicate
tabulate
openai